chore: add recursive check to `is_zero_size`; make `max_serialized_size` a container method
