+++ borsh/src/lib.rs
@@ -91,8 +91,7 @@ pub use de::{from_reader, from_slice};
 pub use schema::BorshSchema;
 #[cfg(feature = "unstable__schema")]
 pub use schema_helpers::{
-    max_serialized_size, schema_container_of, try_from_slice_with_schema, try_to_vec_with_schema,
-    MaxSizeError,
+    max_size_of, schema_container_of, try_from_slice_with_schema, try_to_vec_with_schema,
 };
 pub use ser::helpers::{to_vec, to_writer};
 pub use ser::BorshSerialize;
