+++ borsh/src/schema/container_ext.rs
@@ -1,0 +1,7 @@
+use super::{BorshSchemaContainer, Declaration, Definition, Fields};
+
+// #[allow(unused_imports)]
+// pub(self) use max_size::is_zero_size;
+pub use max_size::MaxSizeError;
+
+mod max_size;
